// web/src/api/models/NotificationEntityType.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */

/**
 * Entity types for notification event catalog - matches backend NotificationEntityType enum
 */
export enum NotificationEntityType {
  JIRA_WEBHOOK = 'JIRA_WEBHOOK',
  ARCGIS_INTEGRATION = 'ARCGIS_INTEGRATION',
  INTEGRATION_CONNECTION = 'INTEGRATION_CONNECTION',
  SITE_CONFIG = 'SITE_CONFIG',
}